server-side configuration of fixed image transformations/resolutions. JSON metadata served per image and global available transformations

+++ make.sh
@@ -9,28 +9,29 @@ buildImage() {
 	docker build -t "$IMAGE_NAME" --rm .
 }
 
-case "$1" in
-	npm)
-		buildJS
-		;;
-	image)
-		buildImage
-		;;
-	'')
-		buildJS &&
-		buildImage
-		;;
-	run)
-		buildJS &&
-		buildImage &&
-		docker run --rm --name webdav-service -p 80:80 "$IMAGE_NAME"	
-		;;
-	work)
-		buildJS &&
-		buildImage &&
-		docker run --rm --name webdav-service -p 80:80 -v "$(pwd)/dist:/var/www/html" "$IMAGE_NAME"	
-		;;
-	*)
-		echo "Usage: $0 [npm|image|run|work]"
-		;;
-esac
+if [ $# -eq 0 ]; then
+	buildJS &&
+	buildImage
+	exit $?
+fi
+
+while [ $# -gt 0 ]; do
+	case "$1" in
+		js)
+			buildJS
+			;;
+		image)
+			buildImage
+			;;
+		run)
+			docker run --rm --name webdav-service -e IMAGE_CROP_27_23=true -e IMAGE_RESIZE_100_100=true -p 80:80 "$IMAGE_NAME"
+			;;
+		work)
+			docker run --rm --name webdav-service -p 80:80 -v "$(pwd)/dist:/var/www/html" "$IMAGE_NAME"
+			;;
+		*)
+			echo "Usage: $0 [js|image|run|work]"
+			;;
+	esac
+	shift
+done
